+++ src/main/java/org/quaerense/laps/service/TaskService.java
@@ -10,7 +10,7 @@ public interface TaskService {
 
     List<Task> getTasks();
 
-    List<Task> getTasksByStatusAndEmployeeId(TaskStatus taskStatus, Integer id);
+    List<Task> getTasksByStatusAndEmployeeId(TaskStatus taskStatus, Long id);
 
     Task getTaskById(Long id);
